CO-1371 Split mass_mailing module and add UTM objects
