+++ wordpress_connector/README.rst
@@ -1,0 +1,19 @@
+.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
+    :alt: License: AGPL-3
+
+Mass Mailing Switzerland
+========================
+
+Tailors mass mailings to Compassion CH needs.
+
+Contributors
+------------
+
+* Emanuel Cino <ecino@compassion.ch>
+* Philippe Heer <heerphilippe@msn.com>
+
+Maintainer
+----------
+
+This module is maintained by `Compassion Switzerland <https://www
+.compassion.ch>`.
