+++ wordpress_connector/views/sponsorship_view.xml
